# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/user/me":{"get":{"operationId":"UserController_me","parameters":[],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/api/user":{"patch":{"operationId":"UserController_updateUser","parameters":[],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/api/user/{idOrEmail}":{"get":{"operationId":"UserController_findOneUser","parameters":[{"name":"idOrEmail","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/api/user/{id}":{"delete":{"operationId":"UserController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/auth/refresh-tokens":{"get":{"operationId":"AuthController_refreshTokens","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/google":{"get":{"operationId":"AuthController_googleAuth","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/google/callback":{"get":{"operationId":"AuthController_googleAuthCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/yandex":{"get":{"operationId":"AuthController_yandexAuth","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/yandex/callback":{"get":{"operationId":"AuthController_yandexAuthCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/success-google":{"get":{"operationId":"AuthController_successGoogle","parameters":[{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/success-yandex":{"get":{"operationId":"AuthController_successYandex","parameters":[{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/pack":{"post":{"operationId":"PackController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePackDto"}}}},"responses":{"201":{"description":""}},"tags":["pack"],"security":[{"bearer":[]}]},"get":{"operationId":"PackController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["pack"],"security":[{"bearer":[]}]}},"/api/pack/{id}/owner":{"get":{"operationId":"PackController_owner","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["pack"],"security":[{"bearer":[]}]}},"/api/pack/{id}":{"get":{"operationId":"PackController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["pack"],"security":[{"bearer":[]}]},"patch":{"operationId":"PackController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePackDto"}}}},"responses":{"200":{"description":""}},"tags":["pack"],"security":[{"bearer":[]}]},"delete":{"operationId":"PackController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["pack"],"security":[{"bearer":[]}]}},"/api/packs/{packId}/card":{"post":{"operationId":"CardController_create","parameters":[{"name":"packId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCardDto"}}}},"responses":{"201":{"description":""}},"tags":["card"],"security":[{"bearer":[]}]},"get":{"operationId":"CardController_findAll","parameters":[{"name":"packId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["card"],"security":[{"bearer":[]}]}},"/api/packs/{packId}/card/{id}":{"get":{"operationId":"CardController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["card"],"security":[{"bearer":[]}]},"patch":{"operationId":"CardController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"packId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCardDto"}}}},"responses":{"200":{"description":""}},"tags":["card"],"security":[{"bearer":[]}]},"delete":{"operationId":"CardController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["card"],"security":[{"bearer":[]}]}},"/api/tag":{"post":{"operationId":"TagController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTagDto"}}}},"responses":{"201":{"description":""}},"tags":["tag"]},"get":{"operationId":"TagController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["tag"]}},"/api/tag/{title}":{"get":{"operationId":"TagController_findByTitle","parameters":[{"name":"title","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["tag"]},"patch":{"operationId":"TagController_update","parameters":[{"name":"title","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTagDto"}}}},"responses":{"200":{"description":""}},"tags":["tag"]},"delete":{"operationId":"TagController_delete","parameters":[{"name":"title","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["tag"]}},"/api/packs/{packId}/session":{"post":{"operationId":"SessionController_create","parameters":[{"name":"packId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSessionDto"}}}},"responses":{"201":{"description":""}},"tags":["session"],"security":[{"bearer":[]}]},"get":{"operationId":"SessionController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["session"],"security":[{"bearer":[]}]}},"/api/packs/{packId}/session/{id}":{"get":{"operationId":"SessionController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["session"],"security":[{"bearer":[]}]},"patch":{"operationId":"SessionController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSessionDto"}}}},"responses":{"200":{"description":""}},"tags":["session"],"security":[{"bearer":[]}]},"delete":{"operationId":"SessionController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["session"],"security":[{"bearer":[]}]}},"/api/sessions/{sessionId}/results":{"post":{"operationId":"ResultController_create","parameters":[{"name":"sessionId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateResultDto"}}}},"responses":{"201":{"description":""}},"tags":["result"],"security":[{"bearer":[]}]},"get":{"operationId":"ResultController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["result"],"security":[{"bearer":[]}]}},"/api/sessions/{sessionId}/results/{id}":{"get":{"operationId":"ResultController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["result"],"security":[{"bearer":[]}]},"patch":{"operationId":"ResultController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateResultDto"}}}},"responses":{"200":{"description":""}},"tags":["result"],"security":[{"bearer":[]}]},"delete":{"operationId":"ResultController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["result"],"security":[{"bearer":[]}]}}},"info":{"title":"Deep Dive Card API","description":"---The DDC API Description---","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"passwordRepeat":{"type":"string"}},"required":["email","password","passwordRepeat"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreatePackDto":{"type":"object","properties":{"title":{"type":"string"},"isPublic":{"type":"boolean"},"tags":{"type":"array","items":{"type":"string"}}},"required":["title"]},"UpdatePackDto":{"type":"object","properties":{"title":{"type":"string"},"isPublic":{"type":"boolean"},"tags":{"type":"array","items":{"type":"string"}}},"required":["title"]},"CreateCardDto":{"type":"object","properties":{"question":{"type":"string"},"packId":{"type":"number"},"answer":{"type":"string"},"hint":{"type":"string"},"difficulty":{"type":"number"}},"required":["question","packId","answer"]},"UpdateCardDto":{"type":"object","properties":{}},"CreateTagDto":{"type":"object","properties":{"title":{"type":"string"}},"required":["title"]},"UpdateTagDto":{"type":"object","properties":{"title":{"type":"string"}},"required":["title"]},"CreateSessionDto":{"type":"object","properties":{"userId":{"type":"string"},"packId":{"type":"number"},"difficulty":{"type":"number"}},"required":["userId","packId"]},"UpdateSessionDto":{"type":"object","properties":{"userId":{"type":"string"},"packId":{"type":"number"},"difficulty":{"type":"number"}}},"CreateResultDto":{"type":"object","properties":{"cardId":{"type":"number"},"userAnswer":{"type":"object"},"usedHint":{"type":"boolean"}},"required":["cardId"]},"UpdateResultDto":{"type":"object","properties":{"cardId":{"type":"number"},"userAnswer":{"type":"object"},"usedHint":{"type":"boolean"}}}}}}